Move ui test utils to src
