+++ ui/src/features/annotator/annotations/annotation-shape.component.test.tsx
@@ -4,7 +4,7 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import { getMockedAnnotation } from '../../../../tests/test-utils/mocked-annotation';
+import { getMockedAnnotation } from '../../../test-utils/mocked-annotation';
 import { Annotation, Polygon, Rect } from '../types';
 import { AnnotationShape } from './annotation-shape.component';
