+++ ui/src/test-utils/mocked-annotation.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Annotation } from '../../src/features/annotator/types';
+import { Annotation } from '../features/annotator/types';
 import { getMockedLabel } from './mocked-labels';
 
 export const getMockedAnnotation = (annotation?: Partial<Annotation>): Annotation => {
