+++ ui/src/test-utils/mocked-labels.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Label } from '../../src/features/annotator/types';
+import { Label } from '../features/annotator/types';
 
 export const getMockedLabel = (label?: Partial<Label>): Label => {
     return {
